Configure paths to work with package.json / settings file from project root
